Refactored some String constants' names in StartsConstants.java and
Rolled back the changes in
starts-core/src/main/java/edu/illinois/starts/asm to its previous
format.
only the Strings which are occurred several times and in different
files, are refactored and moved to StartsConstants.
Strings with multiple occurrance in one class, are replaced with static
final string in the same class.